feat: HTTPServer

+++ DataStructure/Dictionary/Entry.c
@@ -17,12 +17,17 @@
  * @param value_size - size of the value
  * @return Entry
  */
-struct Entry entry_constructor(void *key, int key_size , void *value, int value_size){
+struct Entry entry_constructor(void *key, unsigned long key_size , void *value, unsigned long value_size){
     struct Entry entry;
+
+    // allocating memory
     entry.key = malloc(key_size);
     entry.value = malloc(value_size);
+
+    // copying the data
     memcpy(entry.key,key,key_size);
     memcpy(entry.value,value,value_size);
+
     return entry;
 };
